docs(adjust-type): adjust type of productQuantity

+++ src/main/java/hcmut/pos_system/models/ImportBatch.java
@@ -39,7 +39,7 @@ public class ImportBatch {
 
     @Id
     @Column(name = "ProductQuantity", nullable = false)
-    private Long productQuantity;
+    private Integer productQuantity;
 
     @Id
     @Column(name = "BatchDate", nullable = false)
